second/requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.5
pymongo==4.3.3
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
SQLAlchemy==2.0.4
bson==0.5.10
